pyproject.toml
[tool.poetry]
name = "robotz213"
version = "0.0.1"
description = "Robotz Page"
authors = ["Robotz213 <nicholas@robotz.dev>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.1.0"
flask-jwt-extended = "^4.7.0"
flask-sqlalchemy = "^3.1.1"
flask-bcrypt = "^1.0.1"
flask-form = "^1.0.0"
flask-wtf = "^1.2.2"
flask-mail = "^0.10.0"
selenium = "^4.26.1"
webdriver-manager = "^4.0.2"
flask-marshmallow = "^1.2.1"
python-dotenv = "^1.0.1"
flask-mysqldb = "^2.0.0"
mysql-connector-python = "^9.1.0"
phonenumbers = "^8.13.50"
bootstrap-flask = "^2.4.1"
eventlet = "^0.38.0"
deep-translator = "^1.11.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
